tasks/ppc-drovosec/server-whitebox/requirements.txt
aiortc==1.8.0
fastapi==0.110.2
pillow==10.3.0
numpy==1.26.4
pydantic==2.7.1
